+++ base/stacktraces.jl
@@ -318,7 +318,7 @@ function from(frame::StackFrame, m::Module)
     if finfo isa Core.MethodInstance
         frame_m = finfo.def
         isa(frame_m, Method) && (frame_m = frame_m.module)
-        result = module_name(frame_m) === module_name(m)
+        result = nameof(frame_m) === nameof(m)
     end
 
     return result