+++ stdlib/REPL/src/REPLCompletions.jl
@@ -72,7 +72,7 @@ function complete_symbol(sym, ffunc)
         # We will exclude the results that the user does not want, as well
         # as excluding Main.Main.Main, etc., because that's most likely not what
         # the user wants
-        p = s->(!Base.isdeprecated(mod, s) && s != module_name(mod) && ffunc(mod, s))
+        p = s->(!Base.isdeprecated(mod, s) && s != nameof(mod) && ffunc(mod, s))
         # Looking for a binding in a module
         if mod == context_module
             # Also look in modules we got through `using`