+++ NEWS.md
@@ -964,6 +964,8 @@ Deprecated or removed
 
   * `findin(a, b)` has been deprecated in favor of `findall(occursin(b), a)` ([#24673]).
 
+  * `module_name` has been deprecated in favor of a new, general `nameof` function ([#25622]).
+
   * The module `Random.dSFMT` is renamed `Random.DSFMT` ([#25567]).
 
   * `Random.RandomDevice(unlimited::Bool)` (on non-Windows systems) is deprecated in favor of
@@ -1239,5 +1241,6 @@ Command-line option changes
 [#25532]: https://github.com/JuliaLang/julia/issues/25532
 [#25545]: https://github.com/JuliaLang/julia/issues/25545
 [#25616]: https://github.com/JuliaLang/julia/issues/25616
+[#25622]: https://github.com/JuliaLang/julia/issues/25622
 [#25634]: https://github.com/JuliaLang/julia/issues/25634
 [#25654]: https://github.com/JuliaLang/julia/issues/25654